Move resources to the resourcemanager package (#8144)
