+++ tfplan2cai/converters/google/resources/organization_policy.go
@@ -47,7 +47,7 @@ func GetOrganizationPolicyApiObject(d tpgresource.TerraformResourceData, config 
 	}
 
 	policy := tpgresource.OrgPolicy{
-		Constraint:     canonicalOrgPolicyConstraint(d.Get("constraint").(string)),
+		Constraint:     CanonicalOrgPolicyConstraint(d.Get("constraint").(string)),
 		BooleanPolicy:  expandBooleanOrganizationPolicy(d.Get("boolean_policy").([]interface{})),
 		ListPolicy:     listPolicy,
 		RestoreDefault: restoreDefault,
